kafka interaction wth payment service

+++ src/main/java/com/innowise/orderservice/entity/Item.java
@@ -1,15 +1,11 @@
 package com.innowise.orderservice.entity;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
-import java.util.List;
 import java.util.Objects;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -35,7 +31,7 @@ public class Item {
   private String name;
 
   @Column(name = "price")
-  private long price;
+  private double price;
 
   @Column(name = "deleted")
   private boolean deleted;
